Add integration tests
Add integration tests which exercise all the current endpoints of the
Metadata API.

Also test the possible errors that would be returned to a client.

The addition of these tests highlighted some potential bugs which are
fixed as well. If we pass 'version_id' into metadata it overwrites the
response of the version id that corresponds to the db id. Changing the
when the metadata.data merge happens fixes the issue.

Some of the unit tests required updating as well once these changes were
made to fix the bugs. `created_at` will be added to the metadata
responses by Rails by default.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>
Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>