Remove test for payload signed with wrong key
These integration tests do not currently spin up a Kubernetes cluster
when they run. In this intance of this particular test when a request
is made to the Metadata API with a payload that has been signed with
the wrong key the fb-jwt-auth gem will make two requests to the service
token cache.

Since the seeded public key does not match the private key used to sign
the test request the fb-jwt-auth gem will make a second request to the
service token cache telling it to ignore the redis cache and return the
public key in the Kubernetes config map for the requesting application.

The reason it does this is to cater for times when public keys have been
changed in the config maps. The keys in Redis expire after 10 minutes
anyway but this speeds up the process and allows for zero downtime
public key rotations should they ever be required.

This test will need to be in the acceptance tests where is a Kubernetes
cluster to interact with, therefore we are removing it here.