Preload schemas into JSON::Validator
Local referencing seems to be an issue with the json-schema gem. There
are a number of issues open:

https://github.com/ruby-json-schema/json-schema/issues/190
https://github.com/ruby-json-schema/json-schema/issues/370
https://github.com/ruby-json-schema/json-schema/issues/411

To get around this we are preloading all the schemas into the
JSON::Validator at startup using the _name property as the identifier
for each schema. This should allow the JSON::Validator the ability to
more easily find the necessary referenced schemas.