Add unit test for MetadataSerialiser
This is a basic schema validation for the attributes that the
MetadataSerialiser will be returning when requested.

Also add a fixture for a service. This is only for when creating a
service and as such will not have the service_id or created_at as those
are set by the metadata api before being returned.