Set MetadataPresenter parent controller
The MetadataPresenter service controller uses the `helper` method which
is not available when inheriting from ApplicationController::API.

Set the MetadataPresenter controller to inherit from the Base controller
at startup. This is fine as the API does not mount the engine routes nor
does it use the actual controllers at all.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>