Return 403 instead of 401
A 401 error is returned by a server when a request lacks the correct
authentication credentials. However the 401 usually indicates that a
client can try to validate again.

An example of this is when basic authentication is used to prompot
for a username and a password. The initial request to the server will
response with a 401 which the client will then provide a user with an
opportunity to enter credentials.

This app being an API there is no such mechanism in place. There is no
user to intervene should a request fail to have the valid authentication
credentials. Therefore we should just return a 403 (forbidden).

[More info about 401s here](https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/401)