Return user services without using a join
For the moment we are not going to have collaboration between users. So
to ease the call on the database, as well as avoiding having to extend
the join to make sure only unique services are returned, remove the
table join and just use a straight forward query with created_by.