Add new Sentry gems
SENTRY_DSN environment variable is
looked for by default so need to explicitly set it.

This version of sentry uses the config flag send_default_pii to
decide sensitive data to be sent. This is set to false by default.
This only covers the following:

user ip address
user cookie
request body

Further parameters are filtered in the Sentry configuration.